Try to fix abandoned windows, fix WS switch focus.
f_list[ws] is for a starting point for cycling window focus.  It's not
a starting point for refocusing the correct window for workspaces.
Since our focus is always the topmost window, we refocus the first
client in the workspace that is not hidden.  The c_list[WS] is sorted
with the topmost window first, so the topmost unhidden window should get
focus.  This works well for swapping workspaces.